# api/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_erhan_falcon_api_internal_models.OutOfScope:
    properties:
      created_at:
        type: string
      id:
        type: integer
      kind:
        type: string
      pattern:
        type: string
      program_id:
        type: integer
      source_url:
        type: string
    type: object
  github_com_erhan_falcon_api_internal_models.Program:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      name:
        type: string
      platform:
        type: string
      rules:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handlers.errorResp:
    properties:
      error:
        type: string
    type: object
...
  internal_handlers.oosReq:
    properties:
      kind:
        description: exact|wildcard|regex
        type: string
      pattern:
        type: string
    type: object
  internal_handlers.pageResp:
    properties:
      items:
        items: {}
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  internal_handlers.programReq:
    properties:
      description:
        type: string
      icon_url:
        type: string
      name:
        type: string
      platform:
        type: string
      rules:
        type: string
      slug:
        type: string
    type: object
...
  internal_handlers.runStartResp:
    properties:
      run_id:
        type: integer
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: Bug bounty & continuous recon backend.
  title: Falcon API
  version: 0.1.0
paths:
...
  /api/oos/{id}:
    delete:
      parameters:
      - description: rule id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Delete out-of-scope rule
      tags:
      - oos
  /api/programs:
    get:
      parameters:
      - description: page (1-based)
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.pageResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: List programs
      tags:
      - programs
    post:
      consumes:
      - application/json
      parameters:
      - description: program body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.programReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.Program'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Create program
      tags:
      - programs
  /api/programs/{id}:
    delete:
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Delete program
      tags:
      - programs
    get:
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.Program'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Get program
      tags:
      - programs
    patch:
      consumes:
      - application/json
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      - description: fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.programReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.Program'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Update program
      tags:
      - programs
...
  /api/programs/{id}/oos:
    get:
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.pageResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: List out-of-scope rules
      tags:
      - oos
    post:
      consumes:
      - application/json
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      - description: rule body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.oosReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.OutOfScope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Add out-of-scope rule
      tags:
      - oos
...
  /api/programs/{id}/runs:
    delete:
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Delete all runs for a program
      tags:
      - runs
    get:
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.pageResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: List runs
      tags:
      - runs
...
  /api/runs/{id}:
    delete:
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Delete run
      tags:
      - runs
    get:
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Get run with steps
      tags:
      - runs
...
  /api/scope/{id}/run:
    post:
      parameters:
      - description: scope item id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_handlers.runStartResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Trigger pipeline run
      tags:
      - scope
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token issued by /api/auth/login. Format: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"